+++ src/main/java/sogong/ctf/controller/MemberController.java
@@ -5,9 +5,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import sogong.ctf.domain.Member;
-import sogong.ctf.dto.*;
 import sogong.ctf.dto.request.MemberRequestDTO;
-import sogong.ctf.dto.response.TeamFormDTO;
+import sogong.ctf.dto.request.ProfilePostDTO;
 import sogong.ctf.service.AuthUser;
 import sogong.ctf.service.MemberService;
