--- src/main/java/sogong/ctf/dto/CodeRequestDTO.java
@@ -1,16 +1,0 @@
-package sogong.ctf.dto;
-
-import lombok.Data;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-@RequiredArgsConstructor
-@Getter
-@Setter
-@Data
-public class CodeRequestDTO {
-    private String language;
-    private String code;
-    private Long challengeId;
-}
