+++ src/main/java/sogong/ctf/dto/response/AttemptDTO.java
@@ -1,4 +1,4 @@
-package sogong.ctf.dto;
+package sogong.ctf.dto.response;
 
 import lombok.*;
 import sogong.ctf.domain.CodeStatus;
