+++ src/main/java/sogong/ctf/dto/response/ProfileResponseDTO.java
@@ -1,4 +1,4 @@
-package sogong.ctf.dto;
+package sogong.ctf.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ProfilePostNotPWDTO {
-    private String nickname;
-    private String currentPW;
+public class ProfileResponseDTO {
+    private int attemptCount;
+    private int correctCount;
+    private int incorrectCount;
 }
