+++ src/main/java/sogong/ctf/service/AttemptService.java
@@ -12,8 +12,8 @@ import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sogong.ctf.domain.*;
-import sogong.ctf.dto.AttemptDTO;
-import sogong.ctf.dto.CodeRequestDTO;
+import sogong.ctf.dto.response.AttemptDTO;
+import sogong.ctf.dto.request.CodeRequestDTO;
 import sogong.ctf.repository.AttemptRepository;
 import sogong.ctf.repository.ChallengeRepository;
 import sogong.ctf.repository.MemberRepository;
